feat: implement user authentication and authorization with integration tests

+++ src/controllers/userController.ts
@@ -8,6 +8,9 @@ export const getUserById = async (
   next: NextFunction,
 ) => {
   try {
+    if (req.user.id !== req.params.id) {
+      return res.status(403).json({ message: 'Acesso negado' });
+    }
     const user = await userService.getUserById(req.params.id);
     res.status(status.OK).json(user);
   } catch (err) {
@@ -21,6 +24,10 @@ export const updateUser = async (
   next: NextFunction,
 ) => {
   try {
+    if (req.user.id !== req.params.id) {
+      return res.status(403).json({ message: 'Acesso negado' });
+    }
+
     const user = await userService.updateUser(req.params.id, req.body);
     res.status(status.OK).json(user);
   } catch (err) {
@@ -34,6 +41,10 @@ export const deleteUser = async (
   next: NextFunction,
 ) => {
   try {
+    if (req.user.id !== req.params.id) {
+      return res.status(403).json({ message: 'Acesso negado' });
+    }
+
     await userService.deleteUser(req.params.id);
     res.status(status.ACCEPTED).json({ message: 'Usuário deletado' });
   } catch (err) {
